fix: crash because of attempting to rewrite existing file in snapshot
The `dir` DSL function copies an entire directory from the resource folder to the snapshot. If a
`file` DSL function call is nested inside a `dir` call, the snapshot crashes due to an already
existing file. This fix checks if the file already exists in the snapshot before attempting to make
a copy.